requirements.txt
Flask==3.0.3
gunicorn
flask-mysqldb
python-dotenv
mysqlclient
